requirements.txt
google-cloud-speech==2.17.3
fastapi==0.68.0
uvicorn==0.15.0